+++ src/types.jl
@@ -147,6 +147,12 @@ is the abstract type identifying implementation for CUDA arrays.
 """
 abstract type CudaEngine <: Engine end
 
+# The search direction may be negated, this is indicated by argument `pm` which
+# is either a `+` or a `-`.
+const Plus = typeof(+)
+const Minus = typeof(-)
+const PlusOrMinus = Union{Plus,Minus}
+
 """
     Bound{T,N}